Updated OpenAPI mappings and added reflection transformer.
Refactored `GetOpenApiTypeAndFormat` to use a switch statement and added mappings for additional types. Introduced `DomainPrimitiveOpenApiReflectionTransformer` for schema generation via reflection. Updated test schemas to align with new formats.

+++ tests/AltaSoft.DomainPrimitives.Generator.Tests/Snapshots/DomainPrimitiveGeneratorTest.DecimalValue_GeneratesAllInterfacesAndConverters#OpenApiHelper.g.verified.cs
@@ -40,7 +40,7 @@ public static class OpenApiHelper
             new OpenApiSchema
             {
                 Type = JsonSchemaType.Number,
-                Format = "decimal",
+                Format = "double",
                 Title = "DecimalValue"
             }
         }
